requirements.txt
pandas == 2.2.3
psycopg2 == 2.9.10
sqlalchemy == 2.0.36
python-dotenv == 1.0.1